refactor: fix factor overflow risk and remove redundant ry variable
Agent-Logs-Url: https://github.com/JuKimpfler/GFS_2026_Physik/sessions/fe3c8a52-81c2-464b-a904-7f5f9e295dbb

+++ Pico_code_IDE/Pico_code_IDE.ino
@@ -25,7 +25,7 @@ unsigned long timestamp   = 0;
 int16_t bitCounter        = 0;
 int16_t bits[13];
 bool dataReady            = false;
-int16_t factor            = 300;
+long factor               = 300;
 
 double filteredY[FILTER_SIZE] = {0};
 int16_t filterIndex           = 0;
